Fix templating
The auto-formatting is messing up the templates.  The `%` characters by themselves to specify OCaml code that shouldn't be included in the html need to be in the first column. The auto-formatting keeps preventing that.  So I switched to the alternative syntax.  Unfortunately, the auto-formatting of the html will not align the tags properly....